+++ src/dao/MeganatorDAO.java
@@ -1,4 +1,4 @@
-package model;
+package dao;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -6,6 +6,9 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+
+import model.Question;
+import model.Result;
 
 public class MeganatorDAO {
 	//DB接続用定数
