+++ src/servlet/HomeServlet.java
@@ -1,6 +1,6 @@
 package servlet;
 
-import static model.Constant_text.*;
+import static config.Constant_text.*;
 
 import java.io.IOException;
 import java.util.ArrayList;
